# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Alchemorsel API
  version: 1.0.0
paths:
  /v1/users:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "201":
          description: created
          content:
            text/plain:
              schema:
                type: string
  /v1/tokens:
    post:
      summary: Login and obtain JWT
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: token returned
          content:
            text/plain:
              schema:
                type: string
  /v1/recipes:

    get:
      summary: Search recipes
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: ingredient
          schema:
            type: string
          explode: true
          style: form
        - in: query
          name: tag
          schema:
            type: string
          explode: true
          style: form
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: list of recipes
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object

    post:
      summary: Create a new recipe
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                ingredients:
                  type: array
                  items:
                    type: string
                steps:
                  type: array
                  items:
                    type: string
                tags:
                  type: array
                  items:
                    type: string
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                type: object

  /v1/recipes/{id}/modify:
    post:
      summary: Request a modified recipe
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                prompt:
                  type: string
      responses:
        "201":
          description: created
          content:
            application/json:
              schema:
                type: object